requirements.txt
airbrake==2.1.0
atomicwrites==1.1.5
attrs==18.1.0
boto==2.48.0
boto3==1.7.24
botocore==1.10.24
cachetools==2.1.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
croniter==0.3.22
Django==2.0.5
django-rq==1.1.0
docutils==0.14
factory-boy==2.11.1
Faker==0.8.15
google-api-python-client==1.6.7
google-auth==1.4.1
google-auth-httplib2==0.0.3
gunicorn==19.8.1
httplib2==0.11.3
idna==2.6
jmespath==0.9.3
more-itertools==4.2.0
oauth2client==4.1.2
plaid-python==2.2.1
pluggy==0.6.0
psycopg2==2.7.4
py==1.5.3
pyasn1==0.4.2
pyasn1-modules==0.2.1
pytest==3.6.0
pytest-django==3.2.1
pytest-mock==1.10.0
pytest-socket==0.2.0
python-dateutil==2.7.3
pytz==2018.4
redis==2.10.6
requests==2.18.4
rq==0.10.0
rsa==3.4.2
s3transfer==0.1.13
six==1.11.0
text-unidecode==1.2
uritemplate==3.0.0
urllib3==1.22
